+++ NEWS.md
@@ -1,3 +1,8 @@
+# naryn 2.6.2
+
+* Added support for multiple databases. 
+* Bug fix: allow calling `emr_track.logical.create` when values are a list of length 1. 
+
 # naryn 2.6.1
 
 * Added support for logical tracks.